refactor: eligible recipient logic
